app/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Natural Language SQL Interface Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.13",
    "uvicorn==0.34.3",
    "python-multipart==0.0.20",
    "openai==1.88.0",
    "anthropic==0.54.0",
    "python-dotenv==1.0.1",
    "mangum>=0.20.0",
    "libsql-client>=0.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
